Manage data import.

Manage data import.

Manage data import.

Details

This object should not be initialised directly. Use dettl to create the object.

Import can be run by working with import object returned by dettl or by running top-level functions. Run the import by working with this object if you want to step through the import process stage by stage and inspect the data after each stage.

Public fields

path

Path to directory containing import object

Methods

Public methods

DataImport$new()

DataImport$reload()

DataImport$format()

DataImport$help()

DataImport$extract()

DataImport$transform()

DataImport$load()

DataImport$get_connection()

DataImport$get_extracted_data()

DataImport$get_transformed_data()

DataImport$get_log_table()

DataImport$begin_transaction()

DataImport$rollback_transaction()

DataImport$commit_transaction()

Method new()

Tidy up open connections, rolling back any active transactions

Create DataImport object - shouldn't be called directly. Use dettl to create the object

Usage

DataImport$new(path, db_name)

Arguments

path

Path to directory containing import object

db_name

Database from dettl config to create import object for

Returns

A new `DataImport` object

Method reload()

Reload the objects sources to refresh source code or repair a broken Postgres connection.

Usage

DataImport$reload()

Method format()

Custom formatter for pretty printing object summary.

Usage

DataImport$format(brief = FALSE)

Arguments

brief

If TRUE then print a brief summary.

Method help()

Print help page for the object

Usage

DataImport$help()

Method extract()

Run the extract stage of the data import

Usage

DataImport$extract()

Method transform()

Run the transform stage of the data import

Usage

DataImport$transform()

Method load()

Run the load stage of the data import

Usage

DataImport$load(comment = NULL, dry_run = FALSE, allow_dirty_git = FALSE)

Arguments

comment

Optional comment which will be persisted in the log of the import run in the database.

dry_run

Whether to run in dry run mode. If TRUE then any database changes will be rolled back. Defaults to FALSE.

allow_dirty_git

If TRUE then skips check that the import is up to date with remote git repo. FALSE by default.

Method get_connection()

Get the database connection being used by the import. Used for testing.

Usage

DataImport$get_connection()

Returns

The DBI connection

Method get_extracted_data()

Get the extracted data created by the extract step

Usage

DataImport$get_extracted_data()

Returns

The extracted data

Method get_transformed_data()

Get the transformed data created by the transform step

Usage

DataImport$get_transformed_data()

Returns

The transformed data

Method get_log_table()

Get the name of the log table for a particular import. This is the name of the log table configured in the dettl config

Usage

DataImport$get_log_table()

Returns

Log table name

Method begin_transaction()

Start a transaction

Usage

DataImport$begin_transaction()

Method rollback_transaction()

Rollback a transaction

Usage

DataImport$rollback_transaction()

Method commit_transaction()

Commit a transaction

Usage

DataImport$commit_transaction()

Examples

path <- dettl:::prepare_test_import(
  system.file("examples", "person_information", package = "dettl"),
  system.file("examples", "dettl_config.yml", package = "dettl"))
import_path <- file.path(path, "person_information")

import <- dettl::dettl(import_path, db_name = "test")
import$extract()

#> Running extract /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpckdPvv/file7925d7674d1/person_information

#> Running extract tests R/test_extract.R

#> ✔ |  OK F W S | Context
#> ⠏ |   0       | extract                                                         ⠏ |   0       | extract                                                         ✔ |   1       | extract
#> 
#> ══ Results ═════════════════════════════════════════════════════════════════════
#> [ FAIL 0 | WARN 0 | SKIP 0 | PASS 1 ]
#> 
#> Way to go!

#> All extract tests passed.

import$transform()

#> Running transform /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpckdPvv/file7925d7674d1/person_information

#> Running transform tests R/test_transform.R

#> ✔ |  OK F W S | Context
#> ⠏ |   0       | transform                                                       ⠏ |   0       | transform                                                       ✔ |   1       | transform
#> 
#> ══ Results ═════════════════════════════════════════════════════════════════════
#> [ FAIL 0 | WARN 0 | SKIP 0 | PASS 1 ]

#> All transform tests passed.

import$load()

#> Running load /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpckdPvv/file7925d7674d1/person_information

#> Running load in a transaction:

#> 	- Running test queries before making any changes

#> 	- Running load step

#> 	- Running test queries after making changes

#> 	- Running load tests R/test_load.R

#> ✔ |  OK F W S | Context
#> ⠏ |   0       | load                                                            ⠏ |   0       | load                                                            ✔ |   1       | load
#> 
#> ══ Results ═════════════════════════════════════════════════════════════════════
#> [ FAIL 0 | WARN 0 | SKIP 0 | PASS 1 ]

#> All tests passed, commiting changes to database.